added an info page to include the info from the un-authenticated landing page decoupled the sign in functionality from the un-authenticated landing page tweaked the authenticated landing page
