+++ TheRoundTable-Client/src/Components/NavBar.tsx
@@ -21,15 +21,16 @@ const NavBar = ({ setTheme }: NavProps) => {
 
 
       {/* large screen spread out nav bar */}
-      <div className="hidden xl:flex xl:flex-row xl:justify-end xl:gap-1 xl:items-center xl:mr-2 ">
+      <div className="hidden xl:flex xl:flex-row xl:justify-end xl:gap-2 xl:items-center xl:mr-2 ">
 
+        <a href="/info" className="btn btn-neutral btn-sm font-accent capitalize">info <i className="fa-solid fa-lightbulb" /></a>
         <a href="/calendar" className="btn btn-neutral btn-sm font-accent capitalize">calendar <i className="fa-solid fa-calendar-days" /></a>
         <a className="btn btn-neutral btn-sm font-accent capitalize" target="_blank" href='https://www.google.com/search?q=how+to+play+dnd&oq=how+to+play+dnd'>new to dnd <i className="fa-solid fa-graduation-cap" /></a>
         <a href="/parties" className="btn btn-neutral btn-sm font-accent capitalize">Active Parties<i className="fa-solid fa-users" /></a>
         <a href="/characters" className="btn btn-neutral btn-sm font-accent capitalize">view characters <i className="fa-solid fa-user-group" /></a>
 
         <div className="dropdown dropdown-end ">
-          <label tabIndex={0} className="btn btn-neutral btn-sm font-accent capitalize m-1">themes <i className="fa-solid fa-paintbrush" /></label>
+          <label tabIndex={0} className="btn btn-neutral btn-sm font-accent capitalize ">themes <i className="fa-solid fa-paintbrush" /></label>
           <ul tabIndex={0} className="dropdown-content text-neutral-content menu p-2 shadow bg-neutral ZTOP rounded-box">
             {themes.map((theme, index) => (
               <li onClick={() => { localStorage.setItem('theme', theme); setTheme(theme) }} key={index}><a className="hover:text-neutral-content font-accent capitalize">{theme}</a></li>
@@ -49,10 +50,11 @@ const NavBar = ({ setTheme }: NavProps) => {
             <i className="swap-on fa-solid fa-xmark" />
           </summary>
           <ul className="dropdown-content z-10 menu p-2 shadow bg-neutral rounded-box font-primary">
-            <li> <a href="/calendar" className="btn btn-neutral btn-sm capitalize  hover:text-neutral-content">calendar <i className="fa-solid fa-calendar-days" /> </a> </li>
-            <li> <a className="btn btn-neutral btn-sm capitalize hover:text-neutral-content" target="_blank" href='https://www.ign.com/articles/how-to-play-dungeons-and-dragons'>new to dnd <i className="fa-solid fa-graduation-cap" /> </a> </li>
-            <li> <a href="/parties" className="btn btn-neutral btn-sm capitalize  hover:text-neutral-content">Active Parties<i className="fa-solid fa-users" /> </a> </li>
-            <li> <a href="/characters" className="btn btn-neutral btn-sm capitalize  hover:text-neutral-content">view characters <i className="fa-solid fa-user-group" /> </a> </li>
+            <li className="m-1"> <a href="/info" className="btn btn-neutral btn-sm capitalize  hover:bg-accent hover:text-neutral-content">info <i className="fa-solid fa-lightbulb" /> </a> </li>
+            <li className="m-1"> <a href="/calendar" className="btn btn-neutral btn-sm capitalize  hover:bg-accent hover:text-neutral-content">calendar <i className="fa-solid fa-calendar-days" /> </a> </li>
+            <li className="m-1"> <a className="btn btn-neutral btn-sm capitalize hover:bg-accent hover:text-neutral-content" target="_blank" href='https://www.ign.com/articles/how-to-play-dungeons-and-dragons'>new to dnd <i className="fa-solid fa-graduation-cap" /> </a> </li>
+            <li className="m-1"> <a href="/parties" className="btn btn-neutral btn-sm capitalize  hover:bg-accent hover:text-neutral-content">Active Parties<i className="fa-solid fa-users" /> </a> </li>
+            <li className="m-1"> <a href="/characters" className="btn btn-neutral btn-sm capitalize  hover:bg-accent hover:text-neutral-content">view characters <i className="fa-solid fa-user-group" /> </a> </li>
           </ul>
         </details>
